Adding serviceMode paramter to AddAzureSignalR (#7183)
