+++ src/Aspire.Hosting.Azure.SignalR/PublicAPI.Unshipped.txt
@@ -2,3 +2,7 @@
 *REMOVED*static Aspire.Hosting.AzureSignalRExtensions.AddAzureSignalR(this Aspire.Hosting.IDistributedApplicationBuilder! builder, string! name, System.Action<Aspire.Hosting.ApplicationModel.IResourceBuilder<Aspire.Hosting.ApplicationModel.AzureSignalRResource!>!, Aspire.Hosting.ResourceModuleConstruct!, Azure.Provisioning.SignalR.SignalRService!>? configureResource) -> Aspire.Hosting.ApplicationModel.IResourceBuilder<Aspire.Hosting.ApplicationModel.AzureSignalRResource!>!
 *REMOVED*Aspire.Hosting.ApplicationModel.AzureSignalRResource.AzureSignalRResource(string! name, System.Action<Aspire.Hosting.ResourceModuleConstruct!>! configureConstruct) -> void
 Aspire.Hosting.ApplicationModel.AzureSignalRResource.AzureSignalRResource(string! name, System.Action<Aspire.Hosting.Azure.AzureResourceInfrastructure!>! configureInfrastructure) -> void
+Aspire.Hosting.ApplicationModel.AzureSignalRServiceMode
+Aspire.Hosting.ApplicationModel.AzureSignalRServiceMode.Default = 0 -> Aspire.Hosting.ApplicationModel.AzureSignalRServiceMode
+Aspire.Hosting.ApplicationModel.AzureSignalRServiceMode.Serverless = 1 -> Aspire.Hosting.ApplicationModel.AzureSignalRServiceMode
+static Aspire.Hosting.AzureSignalRExtensions.AddAzureSignalR(this Aspire.Hosting.IDistributedApplicationBuilder! builder, string! name, Aspire.Hosting.ApplicationModel.AzureSignalRServiceMode serviceMode) -> Aspire.Hosting.ApplicationModel.IResourceBuilder<Aspire.Hosting.ApplicationModel.AzureSignalRResource!>!
